Introduce exception blocks
This also splits them out to separate sources.
