+++ MetaBrainz.Common.Json/MetaBrainz.Common.Json.csproj
@@ -16,6 +16,7 @@
 
   <PropertyGroup>
     <AssemblyIsComVisible>false</AssemblyIsComVisible>
+    <DefineConstants>TRACE</DefineConstants>
   </PropertyGroup>
 
   <ItemGroup>
